move station/line id consts to shared
