+++ scripts/generate-map-geometry/ptv/lines/clifton-hill.ts
@@ -1,4 +1,4 @@
-import { JOLIMONT } from "../../../../server/data/station-ids";
+import { JOLIMONT } from "../../../../shared/station-ids";
 import { flexi } from "../../lib/dimensions/flexi-length";
 import { LineBlueprint } from "../../lib/blueprint/line-blueprint";
 import { PathBlueprint } from "../../lib/blueprint/path-blueprint";
